refactor: rename EmbedUtil file
